float for tick price

+++ web_client/src/containers/ticks/index.js
@@ -13,7 +13,7 @@ export class Ticks extends Component {
     if (!ticks) { return null }
     return Object.keys(ticks).map(function(tick, i) {
       return <div key={i}>
-        ticker: {tick}, price: {parseInt(ticks[tick].ask_price, 10).toFixed(2)}
+        ticker: {tick}, price: {parseFloat(ticks[tick].ask_price, 10).toFixed(2)}
       </div>
     })
   }
